Update version of certificate as prereq to block download_certificate
