+++ certificate/classes/admin_setting_upload.php
@@ -18,30 +18,29 @@
 /**
  * Creates an upload form on the settings page
  *
- * @package    mod
- * @subpackage certificate
+ * @package    mod_certificate
  * @copyright  Michael Avelar <michaela@moodlerooms.com>
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
+defined('MOODLE_INTERNAL') || die();
+
 require_once($CFG->libdir.'/adminlib.php');
 
 /**
-* Class extends admin setting class to allow/process an uploaded file
-**/
+ * Class extends admin setting class to allow/process an uploaded file
+ **/
 class mod_certificate_admin_setting_upload extends admin_setting_configtext {
     public function __construct($name, $visiblename, $description, $defaultsetting) {
         parent::__construct($name, $visiblename, $description, $defaultsetting, PARAM_RAW, 50);
     }
 
     function output_html($data, $query='') {
-        // Create a dummy var for this field
+        // Create a dummy var for this field.
         $this->config_write($this->name, '');
 
         return format_admin_setting($this, $this->visiblename,
-                html_writer::link(new moodle_url('/mod/certificate/upload_image.php'), get_string('upload')),
-                $this->description, true, '', null, $query);
+            html_writer::link(new moodle_url('/mod/certificate/upload_image.php'), get_string('upload')),
+            $this->description, true, '', null, $query);
     }
 }
-
-?>
