+++ certificate/deprecatedlib.php
@@ -18,8 +18,7 @@
 /**
  * Deprecated certificate functions.
  *
- * @package    mod
- * @subpackage certificate
+ * @package    mod_certificate
  * @copyright  Mark Nelson <markn@moodle.com>
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
