+++ certificate/settings.php
@@ -18,17 +18,22 @@
 /**
  * Provides some custom settings for the certificate module
  *
- * @package    mod
- * @subpackage certificate
+ * @package    mod_certificate
  * @copyright  Michael Avelar <michaela@moodlerooms.com>
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
 defined('MOODLE_INTERNAL') || die;
 
-require_once($CFG->dirroot.'/mod/certificate/adminsetting.class.php');
+if ($ADMIN->fulltree) {
 
-$settings->add(new mod_certificate_admin_setting_upload('certificate/uploadimage',
-    get_string('uploadimage', 'certificate'), get_string('uploadimagedesc', 'certificate'), ''));
+    $settings->add(new mod_certificate_admin_setting_upload('certificate/uploadimage',
+        get_string('uploadimage', 'mod_certificate'), get_string('uploadimagedesc', 'certificate'), ''));
 
-?>
+    $settings->add(new mod_certificate_admin_setting_font('certificate/fontsans',
+        get_string('fontsans', 'mod_certificate'), get_string('fontsans_desc', 'mod_certificate'), 'freesans'));
+
+    $settings->add(new mod_certificate_admin_setting_font('certificate/fontserif',
+        get_string('fontserif', 'mod_certificate'), get_string('fontserif_desc', 'mod_certificate'), 'freeserif'));
+
+}
