+++ certificate/view.php
@@ -18,16 +18,14 @@
 /**
  * Handles viewing a certificate
  *
- * @package    mod
- * @subpackage certificate
+ * @package    mod_certificate
  * @copyright  Mark Nelson <markn@moodle.com>
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
-//require_once("../../config.php");
-require_once($_SERVER['DOCUMENT_ROOT'] . "/config.php");
+require_once("../../config.php");
+require_once("$CFG->dirroot/mod/certificate/locallib.php");
 require_once("$CFG->dirroot/mod/certificate/deprecatedlib.php");
-require_once("$CFG->dirroot/mod/certificate/lib.php");
 require_once("$CFG->libdir/pdflib.php");
 
 $id = required_param('id', PARAM_INT);    // Course Module ID
@@ -44,7 +42,7 @@ if (!$certificate = $DB->get_record('certificate', array('id'=> $cm->instance)))
     print_error('course module is incorrect');
 }
 
-require_login($course->id, false, $cm);
+require_login($course, false, $cm);
 $context = context_module::instance($cm->id);
 require_capability('mod/certificate:view', $context);
 
@@ -59,9 +57,6 @@ $PAGE->set_context($context);
 $PAGE->set_cm($cm);
 $PAGE->set_title(format_string($certificate->name));
 $PAGE->set_heading(format_string($course->fullname));
-
-// Set the context
-$context = context_module::instance($cm->id);
 
 if (($edit != -1) and $PAGE->user_allowed_editing()) {
      $USER->editing = $edit;
@@ -96,8 +91,8 @@ require("$CFG->dirroot/mod/certificate/type/$certificate->certificatetype/certif
 if (empty($action)) { // Not displaying PDF
     echo $OUTPUT->header();
 
-    /// find out current groups mode
-    groups_print_activity_menu($cm, $CFG->wwwroot . '/mod/certificate/view.php?id=' . $cm->id);
+    $viewurl = new moodle_url('/mod/certificate/view.php', array('id' => $cm->id));
+    groups_print_activity_menu($cm, $viewurl);
     $currentgroup = groups_get_activity_group($cm);
     $groupmode = groups_get_activity_groupmode($cm);
 
@@ -129,27 +124,38 @@ if (empty($action)) { // Not displaying PDF
 
     $link = new moodle_url('/mod/certificate/view.php?id='.$cm->id.'&action=get');
     $button = new single_button($link, $linkname);
-    $button->add_action(new popup_action('click', $link, 'view'.$cm->id, array('height' => 600, 'width' => 800)));
+    if ($certificate->delivery != 1) {
+        $button->add_action(new popup_action('click', $link, 'view' . $cm->id, array('height' => 600, 'width' => 800)));
+    }
 
     echo html_writer::tag('div', $OUTPUT->render($button), array('style' => 'text-align:center'));
     echo $OUTPUT->footer($course);
     exit;
 } else { // Output to pdf
-    // Remove full-stop at the end if it exists, to avoid "..pdf" being created and being filtered by clean_filename
-    $certname = rtrim($certificate->name, '.');
-    $filename = clean_filename("$certname.pdf");
+
+    // No debugging here, sorry.
+    $CFG->debugdisplay = 0;
+    @ini_set('display_errors', '0');
+    @ini_set('log_errors', '1');
+
+    $filename = certificate_get_certificate_filename($certificate, $cm, $course) . '.pdf';
+
+    // PDF contents are now in $file_contents as a string.
+    $filecontents = $pdf->Output('', 'S');
+
     if ($certificate->savecert == 1) {
-        // PDF contents are now in $file_contents as a string
-       $file_contents = $pdf->Output('', 'S');
-       certificate_save_pdf($file_contents, $certrecord->id, $filename, $context->id);
+        certificate_save_pdf($filecontents, $certrecord->id, $filename, $context->id);
     }
+
     if ($certificate->delivery == 0) {
-        $pdf->Output($filename, 'I'); // open in browser
+        // Open in browser.
+        send_file($filecontents, $filename, 0, 0, true, false, 'application/pdf');
     } elseif ($certificate->delivery == 1) {
-        $pdf->Output($filename, 'D'); // force download when create
+        // Force download.
+        send_file($filecontents, $filename, 0, 0, true, true, 'application/pdf');
     } elseif ($certificate->delivery == 2) {
-        certificate_email_student($course, $certificate, $certrecord, $context);
-        $pdf->Output($filename, 'I'); // open in browser
-        $pdf->Output('', 'S'); // send
+        certificate_email_student($course, $certificate, $certrecord, $context, $filecontents, $filename);
+        // Open in browser after sending email.
+        send_file($filecontents, $filename, 0, 0, true, false, 'application/pdf');
     }
 }
